test: Move the `ManagedRosBridge` struct to common
This way the test_bridge module contains only tests and is less bloated.
